Move SecurityUserValueResolver to security-http
