+++ UPGRADE-5.0.md
@@ -17,6 +17,7 @@ SecurityBundle
 --------------
 
  * The `logout_on_user_change` firewall option has been removed.
+ * The `SecurityUserValueResolver` class has been removed.
 
 Translation
 -----------
