+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 -----
 
  * The `logout_on_user_change` firewall option is deprecated and will be removed in 5.0.
+ * deprecated `SecurityUserValueResolver`, use
+   `Symfony\Component\Security\Http\Controller\UserValueResolver` instead.
 
 4.0.0
 -----
