+++ src/Symfony/Component/Security/Http/Controller/UserValueResolver.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\SecurityBundle;
+namespace Symfony\Component\Security\Http\Controller;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Controller\ArgumentValueResolverInterface;
@@ -17,18 +17,13 @@ use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadata;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Http\Controller\UserValueResolver;
-
-@trigger_error(sprintf('The "%s" class is deprecated since version 4.1 and will be removed in 5.0, use "%s" instead.', SecurityUserValueResolver::class, UserValueResolver::class), E_USER_DEPRECATED);
 
 /**
  * Supports the argument type of {@see UserInterface}.
  *
  * @author Iltar van der Berg <kjarli@gmail.com>
- *
- * @deprecated since version 4.1, to be removed in 5.0. Use {@link UserValueResolver} instead
  */
-final class SecurityUserValueResolver implements ArgumentValueResolverInterface
+final class UserValueResolver implements ArgumentValueResolverInterface
 {
     private $tokenStorage;
